requirements.txt
# please make sure you have already a pytorch install that is cuda enabled!
av==11.0.0
einops
flashy>=0.0.1
hydra-core>=1.1
hydra_colorlog
torch==2.6.0 --extra-index-url https://download.pytorch.org/whl/cu124
torchaudio>=2.0.0,<2.6.2 --extra-index-url https://download.pytorch.org/whl/cu124
soundfile
huggingface_hub
tqdm
transformers>=4.48.0  # need Encodec there.
xformers>=0.0.23 --index-url https://download.pytorch.org/whl/cu124
demucs
librosa
soundfile
gradio==5.23.3
gradio[oauth]
pillow
torchmetrics
encodec
protobuf>=3.20.1
filetype
wave
mutagen
fastapi>=0.88.0
pydantic
typer
torchvision>=0.21.0 --extra-index-url https://download.pytorch.org/whl/cu124
#torchtext
pesq
pystoi
julius
spacy==3.7.6
sentencepiece
num2words
numpy<1.26.4
matplotlib